iOS: move status_bar to device/status_bar
