+++ lib/calabash/ios/device/status_bar.rb
