+++ spec/lib/ios/device/status_bar_spec.rb
